+++ modules/net/quic/output.h
@@ -166,4 +166,4 @@ void quic_outq_init(struct sock *sk);
 void quic_outq_free(struct sock *sk);
 
 int quic_outq_flow_control(struct sock *sk, struct quic_stream *stream, u16 bytes, u8 sndblock);
-u32 quic_outq_wspace(struct sock *sk, struct quic_stream *stream);
+u64 quic_outq_wspace(struct sock *sk, struct quic_stream *stream);
